+++ src/js/__tests__/reducers/ping.test.js
@@ -26,5 +26,31 @@ describe('ping reducers', () => {
                 appVersion: 'test'
             }
         )
-    })
+    });
+
+    it('Should handle FETCH_PING_COLLECTION_SUCCEEDED', () => {
+        let action = {
+            type: "FETCH_PING_COLLECTION_SUCCEEDED"
+        }
+        expect(ping(initialState, action)).toEqual(
+            {
+                isLoading: false,
+                apiVersion: 'N.D.',
+                appVersion: 'test',
+            }
+        )
+    });
+
+    it('Should handle FETCH_PING_COLLECTION_FAILED', () => {
+        let action = {
+            type: "FETCH_PING_COLLECTION_FAILED"
+        }
+        expect(ping(initialState, action)).toEqual(
+            {
+                isLoading: false,
+                appVersion: 'test',
+                error: undefined
+            }
+        )
+    });
 })