+++ src/js/actions/LoginAction.ts
@@ -1,0 +1,16 @@
+import GenericCRUDActions from "./GenericCRUDActions";
+import { instance as loginApi } from "../../api/LoginAPI";
+
+export default class LoginAction extends GenericCRUDActions {
+    
+    constructor() {
+
+        console.log("called login action")
+
+        super("login", loginApi);
+        
+    }
+
+}
+
+export const actions = new LoginAction();