+++ src/js/reducers/index.js
@@ -1,10 +1,12 @@
 import { combineReducers } from 'redux';
 import ping from "./app/ping";
 import accounts from "./app/accounts";
+import login from "./app/login";
 
 export default combineReducers(
   {
     accounts,
-    ping
+    ping,
+    login,
   }
 );