+++ src/js/components/Navigation.js
@@ -1,26 +1,24 @@
-import React, { Component } from 'react';
 import { Link } from "react-router-dom";
 import { Row, Col } from "react-bootstrap";
 
-export class Navigation extends Component {
+export const Navigation = (props) => {
 
-  logout() {
+  const logout = (e) => {
+    e.preventDefault();
     localStorage.removeItem('REACT_FE_ACCESS_TOKEN')
   }
-  
-  render() {
-    return (
-      <Row>
-        <Col>
-          <ul className="list-unstyled list-inline">
-            <li className="list-inline-item"><Link to="/">Home</Link></li>
-            <li className="list-inline-item"><Link to="/accounts">Accounts</Link></li>
-            <li className="list-inline-item"><Link to="#" onClick={this.logout}>Logout</Link></li>
-          </ul>
-        </Col>
-      </Row>
-    );
-  }
+
+  return (
+    <Row>
+      <Col>
+        <ul className="list-unstyled list-inline">
+          <li className="list-inline-item"><Link to="/">Home</Link></li>
+          <li className="list-inline-item"><Link to="/accounts">Accounts</Link></li>
+          <li className="list-inline-item"><Link to="#" onClick={logout}>Logout</Link></li>
+        </ul>
+      </Col>
+    </Row>
+  );
 }
 
 export default Navigation;