+++ src/js/actions/LoginActions.ts
@@ -4,10 +4,14 @@ import { instance as loginApi } from "../../api/LoginAPI";
 export default class LoginActions extends GenericCRUDActions {
     
     constructor() {
-
         super("login", loginApi);
-        
     }
+
+    logout(){
+        return (dispatch: Function) => dispatch({type:"LOGOUT"});
+    }
+
+
 
 }